requirements.txt
flask
flask-cors
pandas
numpy
faiss-cpu
gdown
sentence-transformers
scikit-learn
gunicorn
